Rename `to_string()` to `to_string_lossy()`
