+++ opc_comn/src/server/traits/opc_shutdown.rs
@@ -1,0 +1,3 @@
+pub trait OpcShutdown {
+    fn shutdown_request(&self, reason: String) -> windows_core::Result<()>;
+}
